Re-saving record doesn't delete image

+++ app/models/business_items.php
@@ -119,7 +119,7 @@ class BusinessItems extends AppModel {
     }
 
     protected function saveImages() {
-        if(array_key_exists('image', $this->data)) {
+        if(array_key_exists('image', $this->data) && $this->data['image']['error'] == 0) {
             if($image = $this->image()) {
                 Model::load('Images')->delete($image->id);
             }
